sched: synchronously wait for signals
Block all signals we are interested in and use sigwaitinfo() to pause.

This allows to catch signals received before waiting for them (we can
fetch a pending signal with sigwaitinfo() even while that signal is
blocked), as well as getting rid of setting up a signal handler.

Note that children inherit the parent signal mask, so forking commands
(such as mpstat) won't work without unblocking signals first.