provide only real-time signals to the user
Real-time signals are prioritized, and queueable: this means that
multiple instances of the same signal can be queued, and the user can
manage priorities for blocks: the real-time signal with the lowest value
will be handled first.

Note: Linux disposes of 32 real-time signals.

For now, deprecate and ignore the usage of SIGUSR1 and SIGUSR2.

Refs #23