block: separate block execution and update
Get rid of popen() and implement 2 spawn (for fork/exec) and update (for
reading output). This allows to trigger block updates without blocking
the others, and setup the environment and other things in the child, so
we are not messing the parent process.

The block is updated when its process notifies its death.

We reset the timestamp when spawning, since a command can take several
seconds, which will delay its updates.

Also rename the "last_update" field to "timestamp" to avoid confusion.

Refs #23